OhMyArtifacts

Documentation for OhMyArtifacts.

OhMyArtifacts.bind_my_artifact!

OhMyArtifacts.create_my_artifact

OhMyArtifacts.download_my_artifact

OhMyArtifacts.my_artifact_exists

OhMyArtifacts.my_artifact_hash

OhMyArtifacts.my_artifact_path

OhMyArtifacts.my_artifacts_toml!

OhMyArtifacts.unbind_my_artifact!

OhMyArtifacts.@my_artifact

OhMyArtifacts.@my_artifacts_toml!

OhMyArtifacts.bind_my_artifact! — Method

bind_my_artifact!(artifacts_toml::String, name::AbstractString, hash::SHA256; force::Bool = false)

Writes a mapping of name -> hash within the given "Artifacts.toml" file. If force is set to true, this will overwrite a pre-existant mapping, otherwise an error is raised.

OhMyArtifacts.create_my_artifact — Method

create_my_artifact(f::Function)

Creates a new artifact by doing path = f(working_dir), hashing the returned path, and moving it to the artifact store. Returns the identifying hash of this artifact.

f(working_dir) should return an absolute path to a single file at the top level of working_dir.

OhMyArtifacts.download_my_artifact — Method

download_my_artifact([downloadf::Function = Base.download], url, name::AbstractString, artifacts_toml::String;
                     force_bind::Bool = false, downloadf_kwarg...)

Convenient function that do download-create-bind together and return the content hash. Download function downloadf should take two position arguments (i.e. downloadf(url, dest; downloadf_kwarg...)). if force_bind is true, it will overwrite the pre-existant binding.

See also: createmyartifact, bindmyartifact!

OhMyArtifacts.my_artifact_exists — Method

my_artifact_exists(hash::SHA256)

Returns whether or not the given artifact (identified by its SHA256 content hash) exists on-disk.

OhMyArtifacts.my_artifact_hash — Method

my_artifact_hash(name::AbstractString, artifacts_toml::String)

Return the hash found in artifacts_toml with given name, or nothing if not found.

OhMyArtifacts.my_artifact_path — Method

my_artifact_path(hash::SHA256)

Given an artifact (identified by SHA256 content hash), return its installation path. If the artifact does not exist, returns the location it would be installed to.

See also: my_artifact_exists

OhMyArtifacts.my_artifacts_toml! — Method

my_artifacts_toml!(pkg::Union{Module,Base.UUID,Nothing})

Return the path to (or creates) "Artifacts.toml" for the given pkg.

See also: @my_artifacts_toml!

OhMyArtifacts.unbind_my_artifact! — Method

unbind_my_artifact!(artifacts_toml::String, name::AbstractString)

Unbind the given name from the "Artifacts.toml" file. Silently fails if no such binding exists within the file.

OhMyArtifacts.@my_artifact — Macro

@my_artifact op name [hash]

Convenient macro for working with "Artifacts.toml". Requiring a global variable my_artifacts storing the path to "Artifacts.toml" (created by @my_artifacts_toml!) to work correctly.

Usage:

@my_artifact :bind name hash => bind_my_artifact!(my_artifacts, name, hash)

@my_artifact :hash name => my_artifact_hash(name, my_artifacts)

@my_artifact :unbind name => unbind_my_artifact!(my_artifacts, name)

@my_artifact :download name url downloadf kwarg... => download_my_artifact(downloadf, url, name, my_artifacts; kwarg...)

See also: bind_my_artifact!, my_artifact_hash, unbind_my_artifact!, @my_artifacts_toml

OhMyArtifacts.@my_artifacts_toml! — Macro

@my_artifacts_toml!()

Convenience macro that gets/creates a "Artifacts.toml" and parented to the package the calling module belongs to.

See also: my_artifacts_toml!